Add active fonts in Alura Also:
- Add image buttons
- Add Images
- Add text buttons
- Remove old SetSize code
- Remove old AluraRenderType
- Add textured quads back into AluraRenderer (to be replaced by Renderer2D)

+++ Saturn/src/Saturn/Alura/AluraFont.cpp
@@ -196,7 +196,7 @@ namespace Saturn {
 		msdfag::ImmediateAtlasGenerator<S, N, Func, msdfag::BitmapAtlasStorage<Ty, N>> generator( width, height );
 		generator.setThreadCount( 4 );
 		generator.setAttributes( attributes );
-		generator.generate( pThis->Glyphs.data(), pThis->Glyphs.size() );
+		generator.generate( pThis->Glyphs.data(), ( int ) pThis->Glyphs.size() );
 
 		msdf::BitmapConstRef<Ty, N> bitmap = generator.atlasStorage();
